requirements.txt
python-telegram-bot
scikit-learn
numpy
pandas
plotly
python-dotenv
pyyaml